fix(auth): cast verification config env values to integers to prevent resend crash
Resolve verification resend failure caused by type mismatch in auth
verification configuration values loaded from `.env`.

Issue:
- `AUTH_VERIFICATION_EXPIRE` / `AUTH_VERIFICATION_RESEND_COOLDOWN` are read as
  strings from environment.
- Carbon expects numeric values when generating verification URL expiry.
- This caused runtime error:
  `Carbon::rawAddUnit(): Argument #3 ($value) must be of type int|float, string given`.

Changes:
- Updated `config/auth.php` to explicitly cast verification settings:
  - `expire` => `(int) env('AUTH_VERIFICATION_EXPIRE', 60)`
  - `resend_cooldown` => `(int) env('AUTH_VERIFICATION_RESEND_COOLDOWN', 60)`

Result:
- Verification link generation and resend flow no longer crash due to string
  env values.
- Auth email verification tests remain passing.

+++ config/auth.php
@@ -112,8 +112,8 @@ return [
     */
 
     'verification' => [
-        'expire' => env('AUTH_VERIFICATION_EXPIRE', 60),
-        'resend_cooldown' => env('AUTH_VERIFICATION_RESEND_COOLDOWN', 60),
+        'expire' => (int) env('AUTH_VERIFICATION_EXPIRE', 60),
+        'resend_cooldown' => (int) env('AUTH_VERIFICATION_RESEND_COOLDOWN', 60),
     ],
 
     /*
